refactor: rename LegacyHeatmapViewer to HeatmapViewer
