+++ libs/insight-viewer/src/Viewer/HeatmapViewer/HeatmapViewer.styles.ts
