+++ libs/insight-viewer/src/Viewer/HeatmapViewer/HeatmapViewer.types.ts
