+++ libs/insight-viewer/src/Viewer/HeatmapViewer/index.tsx
@@ -5,7 +5,7 @@ import { style } from './HeatmapViewer.styles'
 
 import type { HeatmapViewerProps } from './HeatmapViewer.types'
 
-export function LegacyHeatmapViewer(props: HeatmapViewerProps): ReactElement<HTMLCanvasElement> {
+export function HeatmapViewer(props: HeatmapViewerProps): ReactElement<HTMLCanvasElement> {
   const [canvasRef] = useHeatmapDrawing(props)
 
   return <canvas data-cy-name="heatmap-canvas" ref={canvasRef} style={style} />
