+++ libs/insight-viewer/src/Viewer/HeatmapViewer/useHeatmapDrawing.ts
@@ -4,8 +4,8 @@ import { useEffect, useMemo, useRef, RefObject } from 'react'
 import { HeatmapViewerProps } from './HeatmapViewer.types'
 import { useOverlayContext } from '../../contexts'
 
-import drawHeatmap from '../../utils/LegacyHeatmapViewer/drawHeatmap'
-import getHeatmapImageData from '../../utils/LegacyHeatmapViewer/getHeatmapImageData'
+import drawHeatmap from '../../utils/HeatmapViewer/drawHeatmap'
+import getHeatmapImageData from '../../utils/HeatmapViewer/getHeatmapImageData'
 
 function useHeatmapDrawing({ posMap, threshold }: HeatmapViewerProps): [RefObject<HTMLCanvasElement>] {
   const canvasRef = useRef<HTMLCanvasElement>(null)
