+++ libs/insight-viewer/src/utils/HeatmapViewer/clearHeatmap.ts
