+++ libs/insight-viewer/src/utils/HeatmapViewer/drawHeatmap.ts
