+++ libs/insight-viewer/src/utils/HeatmapViewer/getHeatmapImageData.ts
