+++ libs/insight-viewer/src/utils/HeatmapViewer/getRGBArray.ts
